+++ Worker/UnityMmo/Packages/games.cookiedragon.mmogf.core/MmogfMessages/Core/CommandRequest.cs
@@ -18,9 +18,9 @@ namespace Mmogf.Core
         [Key(3)]
         public int EntityId { get; set; }
         [Key(4)]
-        public int ComponentId { get; set; }
+        public short ComponentId { get; set; }
         [Key(5)]
-        public int CommandId { get; set; }
+        public short CommandId { get; set; }
         [Key(6)]
         public byte[] Payload { get; set; }
     }
